requirements.txt
pillow==11.0.0
qrcode==8.0
